test: Conversion refactoring (#759)
* tests

test

* [feat] RR Create API 1: Add dataflow accessor (#745)

* Add dataflow accessor

* Add enable streaming engine struct tag
Mofe Unmarshall Method to acc2 due ot storage dependency

* Moved dataflow utils to accessor and creates types.go

* Create dataflowutils package

* Renamed testing package for dataflow util

* Added unit tests

* Added empty test files for clients

* Move test to same package

* Add tests for dataflow client

* Update fake for client test

* Make dataflow accessor interface and struct to make it testable

* Remove interface from accessor package

* Add dataflow accessor interface

* Add comments to dataflow client and comments on unit tests

* Move all dataflow dependencies to accessors and remove dataflow utils

* Create dataflow client interface for accessor method to make it unit testable

* tests

tests

* common testing

* change

* change

* changes on comments

* change

* accessor mysql conn

* changes

* change

* change

---------
