+++ conversion/conversion.go
@@ -26,28 +26,17 @@ package conversion
 import (
 	"bufio"
 	"context"
-	"database/sql"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"net"
 	"os"
 	"strings"
 	"sync"
-	"sync/atomic"
-	"syscall"
 	"time"
 
-	"cloud.google.com/go/cloudsqlconn"
 	datastream "cloud.google.com/go/datastream/apiv1"
 	sp "cloud.google.com/go/spanner"
 	database "cloud.google.com/go/spanner/admin/database/apiv1"
-	spanneradmin "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/clients/spanner/admin"
-	storageclient "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/clients/storage"
-	spanneraccessor "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/spanner"
-	storageaccessor "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/storage"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/common/constants"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/common/metrics"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/common/utils"
@@ -57,22 +46,8 @@ import (
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/profiles"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/common"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/csv"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/dynamodb"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/mysql"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/oracle"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/postgres"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/spanner"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/sources/sqlserver"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/spanner/ddl"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/spanner/writer"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/streaming"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	dydb "github.com/aws/aws-sdk-go/service/dynamodb"
-	"github.com/aws/aws-sdk-go/service/dynamodbstreams"
-	mysqldriver "github.com/go-sql-driver/mysql"
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/stdlib"
 	"go.uber.org/zap"
 	adminpb "google.golang.org/genproto/googleapis/spanner/admin/database/v1"
 	"google.golang.org/grpc/metadata"
@@ -99,14 +74,20 @@ func getDatastreamClient(ctx context.Context) *datastream.Client {
 	return datastreamClient
 }
 
+type ConvInterface interface {
+	SchemaConv(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, ioHelper *utils.IOStreams, schemaFromSource SchemaFromSourceInterface) (*internal.Conv, error)
+	DataConv(ctx context.Context, sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, ioHelper *utils.IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool, writeLimit int64, dataFromSource DataFromSourceInterface) (*writer.BatchWriter, error) 
+}
+type ConvImpl struct {}
+
 // SchemaConv performs the schema conversion
 // The SourceProfile param provides the connection details to use the go SQL library.
-func SchemaConv(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, ioHelper *utils.IOStreams) (*internal.Conv, error) {
+func (ci *ConvImpl) SchemaConv(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, ioHelper *utils.IOStreams, schemaFromSource SchemaFromSourceInterface) (*internal.Conv, error) {
 	switch sourceProfile.Driver {
 	case constants.POSTGRES, constants.MYSQL, constants.DYNAMODB, constants.SQLSERVER, constants.ORACLE:
-		return schemaFromDatabase(sourceProfile, targetProfile)
+		return schemaFromSource.schemaFromDatabase(sourceProfile, targetProfile, &GetInfoImpl{}, &common.ProcessSchemaImpl{})
 	case constants.PGDUMP, constants.MYSQLDUMP:
-		return SchemaFromDump(sourceProfile.Driver, targetProfile.Conn.Sp.Dialect, ioHelper)
+		return schemaFromSource.SchemaFromDump(sourceProfile.Driver, targetProfile.Conn.Sp.Dialect, ioHelper, &ProcessDumpByDialectImpl{})
 	default:
 		return nil, fmt.Errorf("schema conversion for driver %s not supported", sourceProfile.Driver)
 	}
@@ -114,7 +95,7 @@ func SchemaConv(sourceProfile profiles.SourceProfile, targetProfile profiles.Tar
 
 // DataConv performs the data conversion
 // The SourceProfile param provides the connection details to use the go SQL library.
-func DataConv(ctx context.Context, sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, ioHelper *utils.IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool, writeLimit int64) (*writer.BatchWriter, error) {
+func (ci *ConvImpl) DataConv(ctx context.Context, sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, ioHelper *utils.IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool, writeLimit int64, dataFromSource DataFromSourceInterface) (*writer.BatchWriter, error) {
 	config := writer.BatchWriterConfig{
 		BytesLimit: 100 * 1000 * 1000,
 		WriteLimit: writeLimit,
@@ -123,619 +104,19 @@ func DataConv(ctx context.Context, sourceProfile profiles.SourceProfile, targetP
 	}
 	switch sourceProfile.Driver {
 	case constants.POSTGRES, constants.MYSQL, constants.DYNAMODB, constants.SQLSERVER, constants.ORACLE:
-		return dataFromDatabase(ctx, sourceProfile, targetProfile, config, conv, client)
+		return dataFromSource.dataFromDatabase(ctx, sourceProfile, targetProfile, config, conv, client, &GetInfoImpl{}, &DataFromDatabaseImpl{}, &SnapshotMigrationImpl{})
 	case constants.PGDUMP, constants.MYSQLDUMP:
 		if conv.SpSchema.CheckInterleaved() {
 			return nil, fmt.Errorf("spanner migration tool does not currently support data conversion from dump files\nif the schema contains interleaved tables. Suggest using direct access to source database\ni.e. using drivers postgres and mysql")
 		}
-		return dataFromDump(sourceProfile.Driver, config, ioHelper, client, conv, dataOnly)
+		return dataFromSource.dataFromDump(sourceProfile.Driver, config, ioHelper, client, conv, dataOnly, &ProcessDumpByDialectImpl{}, &PopulateDataConvImpl{})
 	case constants.CSV:
-		return dataFromCSV(ctx, sourceProfile, targetProfile, config, conv, client)
+		return dataFromSource.dataFromCSV(ctx, sourceProfile, targetProfile, config, conv, client, &PopulateDataConvImpl{}, &csv.CsvImpl{})
 	default:
 		return nil, fmt.Errorf("data conversion for driver %s not supported", sourceProfile.Driver)
 	}
 }
 
-func connectionConfig(sourceProfile profiles.SourceProfile) (interface{}, error) {
-	switch sourceProfile.Driver {
-	// For PG and MYSQL, When called as part of the subcommand flow, host/user/db etc will
-	// never be empty as we error out right during source profile creation. If any of them
-	// are empty, that means this was called through the legacy cmd flow and we create the
-	// string using env vars.
-	case constants.POSTGRES:
-		pgConn := sourceProfile.Conn.Pg
-		if !(pgConn.Host != "" && pgConn.User != "" && pgConn.Db != "") {
-			return profiles.GeneratePGSQLConnectionStr()
-		} else {
-			return profiles.GetSQLConnectionStr(sourceProfile), nil
-		}
-	case constants.MYSQL:
-		// If empty, this is called as part of the legacy mode witih global CLI flags.
-		// When using source-profile mode is used, the sqlConnectionStr is already populated.
-		mysqlConn := sourceProfile.Conn.Mysql
-		if !(mysqlConn.Host != "" && mysqlConn.User != "" && mysqlConn.Db != "") {
-			return profiles.GenerateMYSQLConnectionStr()
-		} else {
-			return profiles.GetSQLConnectionStr(sourceProfile), nil
-		}
-	// For Dynamodb, both legacy and new flows use env vars.
-	case constants.DYNAMODB:
-		return getDynamoDBClientConfig()
-	case constants.SQLSERVER:
-		return profiles.GetSQLConnectionStr(sourceProfile), nil
-	case constants.ORACLE:
-		return profiles.GetSQLConnectionStr(sourceProfile), nil
-	default:
-		return "", fmt.Errorf("driver %s not supported", sourceProfile.Driver)
-	}
-}
-
-func getDbNameFromSQLConnectionStr(driver, sqlConnectionStr string) string {
-	switch driver {
-	case constants.POSTGRES:
-		dbParam := strings.Split(sqlConnectionStr, " ")[4]
-		return strings.Split(dbParam, "=")[1]
-	case constants.MYSQL:
-		return strings.Split(sqlConnectionStr, ")/")[1]
-	case constants.SQLSERVER:
-		splts := strings.Split(sqlConnectionStr, "?database=")
-		return splts[len(splts)-1]
-	case constants.ORACLE:
-		// connection string formate : "oracle://user:password@104.108.154.85:1521/XE"
-		substr := sqlConnectionStr[9:]
-		dbName := strings.Split(substr, ":")[0]
-		return dbName
-	}
-	return ""
-}
-
-func getInfoSchemaForShard(shardConnInfo profiles.DirectConnectionConfig, driver string, targetProfile profiles.TargetProfile) (common.InfoSchema, error) {
-	params := make(map[string]string)
-	params["host"] = shardConnInfo.Host
-	params["user"] = shardConnInfo.User
-	params["dbName"] = shardConnInfo.DbName
-	params["port"] = shardConnInfo.Port
-	params["password"] = shardConnInfo.Password
-	//while adding other sources, a switch-case will be added here on the basis of the driver input param passed.
-	//pased on the driver name, profiles.NewSourceProfileConnection<DBName> will need to be called to create
-	//the source profile information.
-	sourceProfileDialect := &profiles.SourceProfileDialectImpl{}
-	getInfo := utils.GetUtilInfoImpl{}
-	sourceProfileConnectionMySQL, err := sourceProfileDialect.NewSourceProfileConnectionMySQL(params, &getInfo)
-	if err != nil {
-		return nil, fmt.Errorf("cannot parse connection configuration for the primary shard")
-	}
-	sourceProfileConnection := profiles.SourceProfileConnection{Mysql: sourceProfileConnectionMySQL, Ty: profiles.SourceProfileConnectionTypeMySQL}
-	//create a source profile which contains the sourceProfileConnection object for the primary shard
-	//this is done because GetSQLConnectionStr() should not be aware of sharding
-	newSourceProfile := profiles.SourceProfile{Conn: sourceProfileConnection, Ty: profiles.SourceProfileTypeConnection}
-	newSourceProfile.Driver = driver
-	infoSchema, err := GetInfoSchema(newSourceProfile, targetProfile)
-	if err != nil {
-		return nil, err
-	}
-	return infoSchema, nil
-}
-
-func schemaFromDatabase(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile) (*internal.Conv, error) {
-	conv := internal.MakeConv()
-	conv.SpDialect = targetProfile.Conn.Sp.Dialect
-	//handle fetching schema differently for sharded migrations, we only connect to the primary shard to
-	//fetch the schema. We reuse the SourceProfileConnection object for this purpose.
-	var infoSchema common.InfoSchema
-	var err error
-	isSharded := false
-	switch sourceProfile.Ty {
-	case profiles.SourceProfileTypeConfig:
-		isSharded = true
-		//Find Primary Shard Name
-		if sourceProfile.Config.ConfigType == constants.BULK_MIGRATION {
-			schemaSource := sourceProfile.Config.ShardConfigurationBulk.SchemaSource
-			infoSchema, err = getInfoSchemaForShard(schemaSource, sourceProfile.Driver, targetProfile)
-			if err != nil {
-				return conv, err
-			}
-		} else if sourceProfile.Config.ConfigType == constants.DATAFLOW_MIGRATION {
-			schemaSource := sourceProfile.Config.ShardConfigurationDataflow.SchemaSource
-			infoSchema, err = getInfoSchemaForShard(schemaSource, sourceProfile.Driver, targetProfile)
-			if err != nil {
-				return conv, err
-			}
-		} else if sourceProfile.Config.ConfigType == constants.DMS_MIGRATION {
-			// TODO: Define the schema processing logic for DMS migrations here.
-			return conv, fmt.Errorf("dms based migrations are not implemented yet")
-		} else {
-			return conv, fmt.Errorf("unknown type of migration, please select one of bulk, dataflow or dms")
-		}
-	case profiles.SourceProfileTypeCloudSQL:
-		infoSchema, err = GetInfoSchemaFromCloudSQL(sourceProfile, targetProfile)
-		if err != nil {
-			return conv, err
-		}
-
-	default:
-		infoSchema, err = GetInfoSchema(sourceProfile, targetProfile)
-		if err != nil {
-			return conv, err
-		}
-	}
-	additionalSchemaAttributes := internal.AdditionalSchemaAttributes{
-		IsSharded: isSharded,
-	}
-	return conv, common.ProcessSchema(conv, infoSchema, common.DefaultWorkers, additionalSchemaAttributes)
-}
-
-func performSnapshotMigration(config writer.BatchWriterConfig, conv *internal.Conv, client *sp.Client, infoSchema common.InfoSchema, additionalAttributes internal.AdditionalDataAttributes) *writer.BatchWriter {
-	common.SetRowStats(conv, infoSchema)
-	totalRows := conv.Rows()
-	if !conv.Audit.DryRun {
-		conv.Audit.Progress = *internal.NewProgress(totalRows, "Writing data to Spanner", internal.Verbose(), false, int(internal.DataWriteInProgress))
-	}
-	batchWriter := populateDataConv(conv, config, client)
-	common.ProcessData(conv, infoSchema, additionalAttributes)
-	batchWriter.Flush()
-	return batchWriter
-}
-
-func snapshotMigrationHandler(sourceProfile profiles.SourceProfile, config writer.BatchWriterConfig, conv *internal.Conv, client *sp.Client, infoSchema common.InfoSchema) (*writer.BatchWriter, error) {
-	switch sourceProfile.Driver {
-	// Skip snapshot migration via Spanner migration tool for mysql and oracle since dataflow job will job will handle this from backfilled data.
-	case constants.MYSQL, constants.ORACLE, constants.POSTGRES:
-		return &writer.BatchWriter{}, nil
-	case constants.DYNAMODB:
-		return performSnapshotMigration(config, conv, client, infoSchema, internal.AdditionalDataAttributes{ShardId: ""}), nil
-	default:
-		return &writer.BatchWriter{}, fmt.Errorf("streaming migration not supported for driver %s", sourceProfile.Driver)
-	}
-}
-
-func updateShardsWithTuningConfigs(shardedTuningConfig profiles.ShardConfigurationDataflow) {
-	for _, dataShard := range shardedTuningConfig.DataShards {
-		dataShard.DatastreamConfig = shardedTuningConfig.DatastreamConfig
-		dataShard.GcsConfig = shardedTuningConfig.GcsConfig
-		dataShard.DataflowConfig = shardedTuningConfig.DataflowConfig
-	}
-}
-
-func dataFromDatabase(ctx context.Context, sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, config writer.BatchWriterConfig, conv *internal.Conv, client *sp.Client) (*writer.BatchWriter, error) {
-	//handle migrating data for sharded migrations differently
-	//sharded migrations are identified via the config= flag, if that flag is not present
-	//carry on with the existing code path in the else block
-	switch sourceProfile.Ty {
-	case profiles.SourceProfileTypeConfig:
-		////There are three cases to cover here, bulk migrations and sharded migrations (and later DMS)
-		//We provide an if-else based handling for each within the sharded code branch
-		//This will be determined via the configType, which can be "bulk", "dataflow" or "dms"
-		if sourceProfile.Config.ConfigType == constants.BULK_MIGRATION {
-			return dataFromDatabaseForBulkMigration(sourceProfile, targetProfile, config, conv, client)
-		} else if sourceProfile.Config.ConfigType == constants.DATAFLOW_MIGRATION {
-			return dataFromDatabaseForDataflowMigration(targetProfile, ctx, sourceProfile, conv)
-		} else if sourceProfile.Config.ConfigType == constants.DMS_MIGRATION {
-			return dataFromDatabaseForDMSMigration()
-		} else {
-			return nil, fmt.Errorf("configType should be one of 'bulk', 'dataflow' or 'dms'")
-		}
-	default:
-		var infoSchema common.InfoSchema
-		var err error
-		if sourceProfile.Ty == profiles.SourceProfileTypeCloudSQL {
-			infoSchema, err = GetInfoSchemaFromCloudSQL(sourceProfile, targetProfile)
-			if err != nil {
-				return nil, err
-			}
-		} else {
-			infoSchema, err = GetInfoSchema(sourceProfile, targetProfile)
-			if err != nil {
-				return nil, err
-			}
-		}
-		var streamInfo map[string]interface{}
-		// minimal downtime migration for a single shard
-		if sourceProfile.Conn.Streaming {
-			//Generate a job Id
-			migrationJobId := conv.Audit.MigrationRequestId
-			logger.Log.Info(fmt.Sprintf("Creating a migration job with id: %v. This jobId can be used in future commmands (such as cleanup) to refer to this job.\n", migrationJobId))
-			streamInfo, err = infoSchema.StartChangeDataCapture(ctx, conv)
-			if err != nil {
-				return nil, err
-			}
-			bw, err := snapshotMigrationHandler(sourceProfile, config, conv, client, infoSchema)
-			if err != nil {
-				return nil, err
-			}
-			dfOutput, err := infoSchema.StartStreamingMigration(ctx, client, conv, streamInfo)
-			if err != nil {
-				return nil, err
-			}
-			dfJobId := dfOutput.JobID
-			gcloudCmd := dfOutput.GCloudCmd
-			streamingCfg, _ := streamInfo["streamingCfg"].(streaming.StreamingCfg)
-			// Fetch and store the GCS bucket associated with the datastream
-			dsClient := getDatastreamClient(ctx)
-			gcsBucket, gcsDestPrefix, fetchGcsErr := streaming.FetchTargetBucketAndPath(ctx, dsClient, targetProfile.Conn.Sp.Project, streamingCfg.DatastreamCfg.DestinationConnectionConfig)
-			if fetchGcsErr != nil {
-				logger.Log.Info("Could not fetch GCS Bucket, hence Monitoring Dashboard will not contain Metrics for the gcs bucket\n")
-				logger.Log.Debug("Error", zap.Error(fetchGcsErr))
-			}
-
-			// Try to apply lifecycle rule to Datastream destination bucket.
-			gcsConfig := streamingCfg.GcsCfg
-			sc, err := storageclient.NewStorageClientImpl(ctx)
-			if err != nil {
-				return nil, err
-			}
-			sa := storageaccessor.StorageAccessorImpl{}
-			if gcsConfig.TtlInDaysSet {
-				err = sa.ApplyBucketLifecycleDeleteRule(ctx, sc, storageaccessor.StorageBucketMetadata{
-					BucketName:    gcsBucket,
-					Ttl:           gcsConfig.TtlInDays,
-					MatchesPrefix: []string{gcsDestPrefix},
-				})
-				if err != nil {
-					logger.Log.Warn(fmt.Sprintf("\nWARNING: could not update Datastream destination GCS bucket with lifecycle rule, error: %v\n", err))
-					logger.Log.Warn("Please apply the lifecycle rule manually. Continuing...\n")
-				}
-			}
-
-			monitoringResources := metrics.MonitoringMetricsResources{
-				ProjectId:            targetProfile.Conn.Sp.Project,
-				DataflowJobId:        dfOutput.JobID,
-				DatastreamId:         streamingCfg.DatastreamCfg.StreamId,
-				JobMetadataGcsBucket: gcsBucket,
-				PubsubSubscriptionId: streamingCfg.PubsubCfg.SubscriptionId,
-				SpannerInstanceId:    targetProfile.Conn.Sp.Instance,
-				SpannerDatabaseId:    targetProfile.Conn.Sp.Dbname,
-				ShardId:              "",
-				MigrationRequestId:   conv.Audit.MigrationRequestId,
-			}
-			respDash, dashboardErr := monitoringResources.CreateDataflowShardMonitoringDashboard(ctx)
-			var dashboardName string
-			if dashboardErr != nil {
-				dashboardName = ""
-				logger.Log.Info("Creation of the monitoring dashboard failed, please create the dashboard manually")
-				logger.Log.Debug("Error", zap.Error(dashboardErr))
-			} else {
-				dashboardName = strings.Split(respDash.Name, "/")[3]
-				fmt.Printf("Monitoring Dashboard: %+v\n", dashboardName)
-			}
-			// store the generated resources locally in conv, this is used as source of truth for persistence and the UI (should change to persisted values)
-			streaming.StoreGeneratedResources(conv, streamingCfg, dfJobId, gcloudCmd, targetProfile.Conn.Sp.Project, "", internal.GcsResources{BucketName: gcsBucket}, dashboardName)
-			//persist job and shard level data in the metadata db
-			err = streaming.PersistJobDetails(ctx, targetProfile, sourceProfile, conv, migrationJobId, false)
-			if err != nil {
-				logger.Log.Info(fmt.Sprintf("Error storing job details in SMT metadata store...the migration job will still continue as intended. %v", err))
-			} else {
-				//only attempt persisting shard level data if the job level data is persisted
-				err = streaming.PersistResources(ctx, targetProfile, sourceProfile, conv, migrationJobId, constants.DEFAULT_SHARD_ID)
-				if err != nil {
-					logger.Log.Info(fmt.Sprintf("Error storing details for migration job: %s, data shard: %s in SMT metadata store...the migration job will still continue as intended. err = %v\n", migrationJobId, constants.DEFAULT_SHARD_ID, err))
-				}
-			}
-			return bw, nil
-		}
-		//bulk migration for a single shard
-		return performSnapshotMigration(config, conv, client, infoSchema, internal.AdditionalDataAttributes{ShardId: ""}), nil
-	}
-}
-
-// TODO: Define the data processing logic for DMS migrations here.
-func dataFromDatabaseForDMSMigration() (*writer.BatchWriter, error) {
-	return nil, fmt.Errorf("dms configType is not implemented yet, please use one of 'bulk' or 'dataflow'")
-}
-
-// 1. Create batch for each physical shard
-// 2. Create streaming cfg from the config source type.
-// 3. Verify the CFG and update it with SMT defaults
-// 4. Launch the stream for the physical shard
-// 5. Perform streaming migration via dataflow
-func dataFromDatabaseForDataflowMigration(targetProfile profiles.TargetProfile, ctx context.Context, sourceProfile profiles.SourceProfile, conv *internal.Conv) (*writer.BatchWriter, error) {
-	updateShardsWithTuningConfigs(sourceProfile.Config.ShardConfigurationDataflow)
-	//Generate a job Id
-	migrationJobId := conv.Audit.MigrationRequestId
-	fmt.Printf("Creating a migration job with id: %v. This jobId can be used in future commmands (such as cleanup) to refer to this job.\n", migrationJobId)
-	conv.Audit.StreamingStats.ShardToShardResourcesMap = make(map[string]internal.ShardResources)
-	schemaDetails, err := common.GetIncludedSrcTablesFromConv(conv)
-	if err != nil {
-		fmt.Printf("unable to determine tableList from schema, falling back to full database")
-		schemaDetails = map[string]internal.SchemaDetails{}
-	}
-	err = streaming.PersistJobDetails(ctx, targetProfile, sourceProfile, conv, migrationJobId, true)
-	if err != nil {
-		logger.Log.Info(fmt.Sprintf("Error storing job details in SMT metadata store...the migration job will still continue as intended. %v", err))
-	}
-	asyncProcessShards := func(p *profiles.DataShard, mutex *sync.Mutex) common.TaskResult[*profiles.DataShard] {
-		dbNameToShardIdMap := make(map[string]string)
-		for _, l := range p.LogicalShards {
-			dbNameToShardIdMap[l.DbName] = l.LogicalShardId
-		}
-		if p.DataShardId == "" {
-			dataShardId, err := utils.GenerateName("smt-datashard")
-			dataShardId = strings.Replace(dataShardId, "_", "-", -1)
-			if err != nil {
-				return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-			}
-			p.DataShardId = dataShardId
-			fmt.Printf("Data shard id generated: %v\n", p.DataShardId)
-		}
-		streamingCfg := streaming.CreateStreamingConfig(*p)
-		err := streaming.VerifyAndUpdateCfg(&streamingCfg, targetProfile.Conn.Sp.Dbname, schemaDetails)
-		if err != nil {
-			err = fmt.Errorf("failed to process shard: %s, there seems to be an error in the sharding configuration, error: %v", p.DataShardId, err)
-			return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-		}
-		fmt.Printf("Initiating migration for shard: %v\n", p.DataShardId)
-		pubsubCfg, err := streaming.CreatePubsubResources(ctx, targetProfile.Conn.Sp.Project, streamingCfg.DatastreamCfg.DestinationConnectionConfig, targetProfile.Conn.Sp.Dbname)
-		if err != nil {
-			return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-		}
-		streamingCfg.PubsubCfg = *pubsubCfg
-		err = streaming.LaunchStream(ctx, sourceProfile, p.LogicalShards, targetProfile.Conn.Sp.Project, streamingCfg.DatastreamCfg)
-		if err != nil {
-			return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-		}
-		streamingCfg.DataflowCfg.DbNameToShardIdMap = dbNameToShardIdMap
-		dfOutput, err := streaming.StartDataflow(ctx, targetProfile, streamingCfg, conv)
-		if err != nil {
-			return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-		}
-		// store the generated resources locally in conv, this is used as source of truth for persistence and the UI (should change to persisted values)
-
-		// Fetch and store the GCS bucket associated with the datastream
-		dsClient := getDatastreamClient(ctx)
-		gcsBucket, gcsDestPrefix, fetchGcsErr := streaming.FetchTargetBucketAndPath(ctx, dsClient, targetProfile.Conn.Sp.Project, streamingCfg.DatastreamCfg.DestinationConnectionConfig)
-		if fetchGcsErr != nil {
-			logger.Log.Info(fmt.Sprintf("Could not fetch GCS Bucket for Shard %s hence Monitoring Dashboard will not contain Metrics for the gcs bucket\n", p.DataShardId))
-			logger.Log.Debug("Error", zap.Error(fetchGcsErr))
-		}
-
-		// Try to apply lifecycle rule to Datastream destination bucket.
-		gcsConfig := streamingCfg.GcsCfg
-		sc, err := storageclient.NewStorageClientImpl(ctx)
-		if err != nil {
-			return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-		}
-		sa := storageaccessor.StorageAccessorImpl{}
-		if gcsConfig.TtlInDaysSet {
-			err = sa.ApplyBucketLifecycleDeleteRule(ctx, sc, storageaccessor.StorageBucketMetadata{
-				BucketName:    gcsBucket,
-				Ttl:           gcsConfig.TtlInDays,
-				MatchesPrefix: []string{gcsDestPrefix},
-			})
-			if err != nil {
-				logger.Log.Warn(fmt.Sprintf("\nWARNING: could not update Datastream destination GCS bucket with lifecycle rule, error: %v\n", err))
-				logger.Log.Warn("Please apply the lifecycle rule manually. Continuing...\n")
-			}
-		}
-
-		// create monitoring dashboard for a single shard
-		monitoringResources := metrics.MonitoringMetricsResources{
-			ProjectId:            targetProfile.Conn.Sp.Project,
-			DataflowJobId:        dfOutput.JobID,
-			DatastreamId:         streamingCfg.DatastreamCfg.StreamId,
-			JobMetadataGcsBucket: gcsBucket,
-			PubsubSubscriptionId: streamingCfg.PubsubCfg.SubscriptionId,
-			SpannerInstanceId:    targetProfile.Conn.Sp.Instance,
-			SpannerDatabaseId:    targetProfile.Conn.Sp.Dbname,
-			ShardId:              p.DataShardId,
-			MigrationRequestId:   conv.Audit.MigrationRequestId,
-		}
-		respDash, dashboardErr := monitoringResources.CreateDataflowShardMonitoringDashboard(ctx)
-		var dashboardName string
-		if dashboardErr != nil {
-			dashboardName = ""
-			logger.Log.Info(fmt.Sprintf("Creation of the monitoring dashboard for shard %s failed, please create the dashboard manually\n", p.DataShardId))
-			logger.Log.Debug("Error", zap.Error(dashboardErr))
-		} else {
-			dashboardName = strings.Split(respDash.Name, "/")[3]
-			fmt.Printf("Monitoring Dashboard for shard %v: %+v\n", p.DataShardId, dashboardName)
-		}
-		streaming.StoreGeneratedResources(conv, streamingCfg, dfOutput.JobID, dfOutput.GCloudCmd, targetProfile.Conn.Sp.Project, p.DataShardId, internal.GcsResources{BucketName: gcsBucket}, dashboardName)
-		//persist the generated resources in a metadata db
-		err = streaming.PersistResources(ctx, targetProfile, sourceProfile, conv, migrationJobId, p.DataShardId)
-		if err != nil {
-			fmt.Printf("Error storing generated resources in SMT metadata store for dataShardId: %s...the migration job will still continue as intended, error: %v\n", p.DataShardId, err)
-		}
-		return common.TaskResult[*profiles.DataShard]{Result: p, Err: err}
-	}
-	_, err = common.RunParallelTasks(sourceProfile.Config.ShardConfigurationDataflow.DataShards, 20, asyncProcessShards, true)
-	if err != nil {
-		return nil, fmt.Errorf("unable to start minimal downtime migrations: %v", err)
-	}
-
-	// create monitoring aggregated dashboard for sharded migration
-	aggMonitoringResources := metrics.MonitoringMetricsResources{
-		ProjectId:                targetProfile.Conn.Sp.Project,
-		SpannerInstanceId:        targetProfile.Conn.Sp.Instance,
-		SpannerDatabaseId:        targetProfile.Conn.Sp.Dbname,
-		ShardToShardResourcesMap: conv.Audit.StreamingStats.ShardToShardResourcesMap,
-		MigrationRequestId:       conv.Audit.MigrationRequestId,
-	}
-	aggRespDash, dashboardErr := aggMonitoringResources.CreateDataflowAggMonitoringDashboard(ctx)
-	if dashboardErr != nil {
-		logger.Log.Error(fmt.Sprintf("Creation of the aggregated monitoring dashboard failed, please create the dashboard manually\n error=%v\n", dashboardErr))
-	} else {
-		fmt.Printf("Aggregated Monitoring Dashboard: %+v\n", strings.Split(aggRespDash.Name, "/")[3])
-		conv.Audit.StreamingStats.AggMonitoringResources = internal.MonitoringResources{DashboardName: strings.Split(aggRespDash.Name, "/")[3]}
-	}
-	err = streaming.PersistAggregateMonitoringResources(ctx, targetProfile, sourceProfile, conv, migrationJobId)
-	if err != nil {
-		logger.Log.Info(fmt.Sprintf("Unable to store aggregated monitoring dashboard in metadata database\n error=%v\n", err))
-	} else {
-		logger.Log.Debug("Aggregate monitoring resources stored successfully.\n")
-	}
-	return &writer.BatchWriter{}, nil
-}
-
-// 1. Migrate the data from the data shards, the schema shard needs to be specified here again.
-// 2. Create a connection profile object for it
-// 3. Perform a snapshot migration for the shard
-// 4. Once all shard migrations are complete, return the batch writer object
-func dataFromDatabaseForBulkMigration(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, config writer.BatchWriterConfig, conv *internal.Conv, client *sp.Client) (*writer.BatchWriter, error) {
-	var bw *writer.BatchWriter
-	for _, dataShard := range sourceProfile.Config.ShardConfigurationBulk.DataShards {
-
-		fmt.Printf("Initiating migration for shard: %v\n", dataShard.DbName)
-		infoSchema, err := getInfoSchemaForShard(dataShard, sourceProfile.Driver, targetProfile)
-		if err != nil {
-			return nil, err
-		}
-		additionalDataAttributes := internal.AdditionalDataAttributes{
-			ShardId: dataShard.DataShardId,
-		}
-		bw = performSnapshotMigration(config, conv, client, infoSchema, additionalDataAttributes)
-	}
-
-	return bw, nil
-}
-
-func getDynamoDBClientConfig() (*aws.Config, error) {
-	cfg := aws.Config{}
-	endpointOverride := os.Getenv("DYNAMODB_ENDPOINT_OVERRIDE")
-	if endpointOverride != "" {
-		cfg.Endpoint = aws.String(endpointOverride)
-	}
-	return &cfg, nil
-}
-
-func SchemaFromDump(driver string, spDialect string, ioHelper *utils.IOStreams) (*internal.Conv, error) {
-	f, n, err := getSeekable(ioHelper.In)
-	if err != nil {
-		utils.PrintSeekError(driver, err, ioHelper.Out)
-		return nil, fmt.Errorf("can't get seekable input file")
-	}
-	ioHelper.SeekableIn = f
-	ioHelper.BytesRead = n
-	conv := internal.MakeConv()
-	conv.SpDialect = spDialect
-	p := internal.NewProgress(n, "Generating schema", internal.Verbose(), false, int(internal.SchemaCreationInProgress))
-	r := internal.NewReader(bufio.NewReader(f), p)
-	conv.SetSchemaMode() // Build schema and ignore data in dump.
-	conv.SetDataSink(nil)
-	err = ProcessDump(driver, conv, r)
-	if err != nil {
-		fmt.Fprintf(ioHelper.Out, "Failed to parse the data file: %v", err)
-		return nil, fmt.Errorf("failed to parse the data file")
-	}
-	p.Done()
-	return conv, nil
-}
-
-func dataFromDump(driver string, config writer.BatchWriterConfig, ioHelper *utils.IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool) (*writer.BatchWriter, error) {
-	// TODO: refactor of the way we handle getSeekable
-	// to avoid the code duplication here
-	if !dataOnly {
-		_, err := ioHelper.SeekableIn.Seek(0, 0)
-		if err != nil {
-			fmt.Printf("\nCan't seek to start of file (preparation for second pass): %v\n", err)
-			return nil, fmt.Errorf("can't seek to start of file")
-		}
-	} else {
-		// Note: input file is kept seekable to plan for future
-		// changes in showing progress for data migration.
-		f, n, err := getSeekable(ioHelper.In)
-		if err != nil {
-			utils.PrintSeekError(driver, err, ioHelper.Out)
-			return nil, fmt.Errorf("can't get seekable input file")
-		}
-		ioHelper.SeekableIn = f
-		ioHelper.BytesRead = n
-	}
-	totalRows := conv.Rows()
-
-	conv.Audit.Progress = *internal.NewProgress(totalRows, "Writing data to Spanner", internal.Verbose(), false, int(internal.DataWriteInProgress))
-	r := internal.NewReader(bufio.NewReader(ioHelper.SeekableIn), nil)
-	batchWriter := populateDataConv(conv, config, client)
-	ProcessDump(driver, conv, r)
-	batchWriter.Flush()
-	conv.Audit.Progress.Done()
-
-	return batchWriter, nil
-}
-
-func dataFromCSV(ctx context.Context, sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile, config writer.BatchWriterConfig, conv *internal.Conv, client *sp.Client) (*writer.BatchWriter, error) {
-	if targetProfile.Conn.Sp.Dbname == "" {
-		return nil, fmt.Errorf("dbName is mandatory in target-profile for csv source")
-	}
-	conv.SpDialect = targetProfile.Conn.Sp.Dialect
-	dialect, err := targetProfile.FetchTargetDialect(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not fetch dialect: %v", err)
-	}
-	if strings.ToLower(dialect) != constants.DIALECT_POSTGRESQL {
-		dialect = constants.DIALECT_GOOGLESQL
-	}
-
-	if dialect != conv.SpDialect {
-		return nil, fmt.Errorf("dialect specified in target profile does not match spanner dialect")
-	}
-
-	delimiterStr := sourceProfile.Csv.Delimiter
-	if len(delimiterStr) != 1 {
-		return nil, fmt.Errorf("delimiter should only be a single character long, found '%s'", delimiterStr)
-	}
-
-	delimiter := rune(delimiterStr[0])
-
-	err = utils.ReadSpannerSchema(ctx, conv, client)
-	if err != nil {
-		return nil, fmt.Errorf("error trying to read and convert spanner schema: %v", err)
-	}
-
-	tables, err := csv.GetCSVFiles(conv, sourceProfile)
-	if err != nil {
-		return nil, fmt.Errorf("error finding csv files: %v", err)
-	}
-
-	// Find the number of rows in each csv file for generating stats.
-	err = csv.SetRowStats(conv, tables, delimiter)
-	if err != nil {
-		return nil, err
-	}
-
-	totalRows := conv.Rows()
-	conv.Audit.Progress = *internal.NewProgress(totalRows, "Writing data to Spanner", internal.Verbose(), false, int(internal.DataWriteInProgress))
-	batchWriter := populateDataConv(conv, config, client)
-	err = csv.ProcessCSV(conv, tables, sourceProfile.Csv.NullStr, delimiter)
-	if err != nil {
-		return nil, fmt.Errorf("can't process csv: %v", err)
-	}
-	batchWriter.Flush()
-	conv.Audit.Progress.Done()
-	return batchWriter, nil
-}
-
-func populateDataConv(conv *internal.Conv, config writer.BatchWriterConfig, client *sp.Client) *writer.BatchWriter {
-	rows := int64(0)
-	config.Write = func(m []*sp.Mutation) error {
-		ctx := context.Background()
-		if !conv.Audit.SkipMetricsPopulation {
-			migrationData := metrics.GetMigrationData(conv, "", constants.DataConv)
-			serializedMigrationData, _ := proto.Marshal(migrationData)
-			migrationMetadataValue := base64.StdEncoding.EncodeToString(serializedMigrationData)
-			ctx = metadata.AppendToOutgoingContext(context.Background(), constants.MigrationMetadataKey, migrationMetadataValue)
-		}
-		_, err := client.Apply(ctx, m)
-		if err != nil {
-			return err
-		}
-		atomic.AddInt64(&rows, int64(len(m)))
-		conv.Audit.Progress.MaybeReport(atomic.LoadInt64(&rows))
-		return nil
-	}
-	batchWriter := writer.NewBatchWriter(config)
-	conv.SetDataMode()
-	if !conv.Audit.DryRun {
-		conv.SetDataSink(
-			func(table string, cols []string, vals []interface{}) {
-				batchWriter.AddRow(table, cols, vals)
-			})
-		conv.DataFlush = func() {
-			batchWriter.Flush()
-		}
-	}
-
-	return batchWriter
-}
 
 // Report generates a report of schema and data conversion.
 func Report(driver string, badWrites map[string]int64, BytesRead int64, banner string, conv *internal.Conv, reportFileName string, dbName string, out *os.File) {
@@ -786,88 +167,6 @@ func Report(driver string, badWrites map[string]int64, BytesRead int64, banner s
 		fmt.Fprint(out, structuredReport.Summary.Text)
 		fmt.Fprintf(out, "See file '%s' for details of the schema and data conversions.\n", reportFileName)
 	}
-}
-
-// getSeekable returns a seekable file (with same content as f) and the size of the content (in bytes).
-func getSeekable(f *os.File) (*os.File, int64, error) {
-	_, err := f.Seek(0, 0)
-	if err == nil { // Stdin is seekable, let's just use that. This happens when you run 'cmd < file'.
-		n, err := utils.GetFileSize(f)
-		return f, n, err
-	}
-	internal.VerbosePrintln("Creating a tmp file with a copy of stdin because stdin is not seekable.")
-	logger.Log.Debug("Creating a tmp file with a copy of stdin because stdin is not seekable.")
-
-	// Create file in os.TempDir. Its not clear this is a good idea e.g. if the
-	// pg_dump/mysqldump output is large (tens of GBs) and os.TempDir points to a directory
-	// (such as /tmp) that's configured with a small amount of disk space.
-	// To workaround such limits on Unix, set $TMPDIR to a directory with lots
-	// of disk space.
-	fcopy, err := ioutil.TempFile("", "spanner-migration-tool.data")
-	if err != nil {
-		return nil, 0, err
-	}
-	syscall.Unlink(fcopy.Name()) // File will be deleted when this process exits.
-	_, err = io.Copy(fcopy, f)
-	if err != nil {
-		return nil, 0, fmt.Errorf("can't write stdin to tmp file: %w", err)
-	}
-	_, err = fcopy.Seek(0, 0)
-	if err != nil {
-		return nil, 0, fmt.Errorf("can't reset file offset: %w", err)
-	}
-	n, _ := utils.GetFileSize(fcopy)
-	return fcopy, n, nil
-}
-
-// VerifyDb checks whether the db exists and if it does, verifies if the schema is what we currently support.
-func VerifyDb(ctx context.Context, adminClient *database.DatabaseAdminClient, dbURI string) (dbExists bool, err error) {
-	adminClientImpl, err := spanneradmin.NewAdminClientImpl(ctx)
-	if err != nil {
-		return dbExists, err
-	}
-	spA := spanneraccessor.SpannerAccessorImpl{}
-	dbExists, err = spA.CheckExistingDb(ctx, adminClientImpl, dbURI)
-	if err != nil {
-		return dbExists, err
-	}
-	if dbExists {
-		err = ValidateDDL(ctx, adminClient, dbURI)
-	}
-	return dbExists, err
-}
-
-// ValidateTables validates that all the tables in the database are empty.
-// It returns the name of the first non-empty table if found, and an empty string otherwise.
-func ValidateTables(ctx context.Context, client *sp.Client, spDialect string) (string, error) {
-	infoSchema := spanner.InfoSchemaImpl{Client: client, Ctx: ctx, SpDialect: spDialect}
-	tables, err := infoSchema.GetTables()
-	if err != nil {
-		return "", err
-	}
-	for _, table := range tables {
-		count, err := infoSchema.GetRowCount(table)
-		if err != nil {
-			return "", err
-		}
-		if count != 0 {
-			return table.Name, nil
-		}
-	}
-	return "", nil
-}
-
-// ValidateDDL verifies if an existing DB's ddl follows what is supported by Spanner migration tool. Currently,
-// we only support empty schema when db already exists.
-func ValidateDDL(ctx context.Context, adminClient *database.DatabaseAdminClient, dbURI string) error {
-	dbDdl, err := adminClient.GetDatabaseDdl(ctx, &adminpb.GetDatabaseDdlRequest{Database: dbURI})
-	if err != nil {
-		return fmt.Errorf("can't fetch database ddl: %v", err)
-	}
-	if len(dbDdl.Statements) != 0 {
-		return fmt.Errorf("spanner migration tool supports writing to existing databases only if they have an empty schema")
-	}
-	return nil
 }
 
 // CreatesOrUpdatesDatabase updates an existing Spanner database or creates a new one if one does not exist.
@@ -1055,375 +354,4 @@ admin quota limit by spreading the FK creation requests over time.`)
 	conv.Audit.Progress.UpdateProgress("Foreign key update complete.", 100, internal.ForeignKeyUpdateComplete)
 	conv.Audit.Progress.Done()
 	return nil
-}
-
-// WriteSchemaFile writes DDL statements in a file. It includes CREATE TABLE
-// statements and ALTER TABLE statements to add foreign keys.
-// The parameter name should end with a .txt.
-func WriteSchemaFile(conv *internal.Conv, now time.Time, name string, out *os.File, driver string) {
-	f, err := os.Create(name)
-	if err != nil {
-		fmt.Fprintf(out, "Can't create schema file %s: %v\n", name, err)
-		return
-	}
-
-	// The schema file we write out below is optimized for reading. It includes comments, foreign keys
-	// and doesn't add backticks around table and column names. This file is
-	// intended for explanatory and documentation purposes, and is not strictly
-	// legal Cloud Spanner DDL (Cloud Spanner doesn't currently support comments).
-	spDDL := conv.SpSchema.GetDDL(ddl.Config{Comments: true, ProtectIds: false, Tables: true, ForeignKeys: true, SpDialect: conv.SpDialect, Source: driver})
-	if len(spDDL) == 0 {
-		spDDL = []string{"\n-- Schema is empty -- no tables found\n"}
-	}
-	l := []string{
-		fmt.Sprintf("-- Schema generated %s\n", now.Format("2006-01-02 15:04:05")),
-		strings.Join(spDDL, ";\n\n"),
-		"\n",
-	}
-	if _, err := f.WriteString(strings.Join(l, "")); err != nil {
-		fmt.Fprintf(out, "Can't write out schema file: %v\n", err)
-		return
-	}
-	fmt.Fprintf(out, "Wrote schema to file '%s'.\n", name)
-
-	// Convert <file_name>.<ext> to <file_name>.ddl.<ext>.
-	nameSplit := strings.Split(name, ".")
-	nameSplit = append(nameSplit[:len(nameSplit)-1], "ddl", nameSplit[len(nameSplit)-1])
-	name = strings.Join(nameSplit, ".")
-	f, err = os.Create(name)
-	if err != nil {
-		fmt.Fprintf(out, "Can't create legal schema ddl file %s: %v\n", name, err)
-		return
-	}
-
-	// We change 'Comments' to false and 'ProtectIds' to true below to write out a
-	// schema file that is a legal Cloud Spanner DDL.
-	spDDL = conv.SpSchema.GetDDL(ddl.Config{Comments: false, ProtectIds: true, Tables: true, ForeignKeys: true, SpDialect: conv.SpDialect, Source: driver})
-	if len(spDDL) == 0 {
-		spDDL = []string{"\n-- Schema is empty -- no tables found\n"}
-	}
-	l = []string{
-		strings.Join(spDDL, ";\n\n"),
-		"\n",
-	}
-	if _, err = f.WriteString(strings.Join(l, "")); err != nil {
-		fmt.Fprintf(out, "Can't write out legal schema ddl file: %v\n", err)
-		return
-	}
-	fmt.Fprintf(out, "Wrote legal schema ddl to file '%s'.\n", name)
-}
-
-// WriteSessionFile writes conv struct to a file in JSON format.
-func WriteSessionFile(conv *internal.Conv, name string, out *os.File) {
-	f, err := os.Create(name)
-	if err != nil {
-		fmt.Fprintf(out, "Can't create session file %s: %v\n", name, err)
-		return
-	}
-	// Session file will basically contain 'conv' struct in JSON format.
-	// It contains all the information for schema and data conversion state.
-	convJSON, err := json.MarshalIndent(conv, "", " ")
-	if err != nil {
-		fmt.Fprintf(out, "Can't encode session state to JSON: %v\n", err)
-		return
-	}
-	if _, err := f.Write(convJSON); err != nil {
-		fmt.Fprintf(out, "Can't write out session file: %v\n", err)
-		return
-	}
-	fmt.Fprintf(out, "Wrote session to file '%s'.\n", name)
-}
-
-// WriteConvGeneratedFiles creates a directory labeled downloads with the current timestamp
-// where it writes the sessionfile, report summary and DDLs then returns the directory where it writes.
-func WriteConvGeneratedFiles(conv *internal.Conv, dbName string, driver string, BytesRead int64, out *os.File) (string, error) {
-	now := time.Now()
-	dirPath := "spanner_migration_tool_output/" + dbName + "/"
-	err := os.MkdirAll(dirPath, os.ModePerm)
-	if err != nil {
-		fmt.Fprintf(out, "Can't create directory %s: %v\n", dirPath, err)
-		return "", err
-	}
-	schemaFileName := dirPath + dbName + "_schema.txt"
-	WriteSchemaFile(conv, now, schemaFileName, out, driver)
-	reportFileName := dirPath + dbName
-	Report(driver, nil, BytesRead, "", conv, reportFileName, dbName, out)
-	sessionFileName := dirPath + dbName + ".session.json"
-	WriteSessionFile(conv, sessionFileName, out)
-	return dirPath, nil
-}
-
-// ReadSessionFile reads a session JSON file and
-// unmarshal it's content into *internal.Conv.
-func ReadSessionFile(conv *internal.Conv, sessionJSON string) error {
-	s, err := ioutil.ReadFile(sessionJSON)
-	if err != nil {
-		return err
-	}
-	err = json.Unmarshal(s, &conv)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// WriteBadData prints summary stats about bad rows and writes detailed info
-// to file 'name'.
-func WriteBadData(bw *writer.BatchWriter, conv *internal.Conv, banner, name string, out *os.File) {
-	badConversions := conv.BadRows()
-	badWrites := utils.SumMapValues(bw.DroppedRowsByTable())
-
-	badDataStreaming := int64(0)
-	if conv.Audit.StreamingStats.Streaming {
-		badDataStreaming = getBadStreamingDataCount(conv)
-	}
-
-	if badConversions == 0 && badWrites == 0 && badDataStreaming == 0 {
-		os.Remove(name) // Cleanup bad-data file from previous run.
-		return
-	}
-	f, err := os.Create(name)
-	if err != nil {
-		fmt.Fprintf(out, "Can't write out bad data file: %v\n", err)
-		return
-	}
-	f.WriteString(banner)
-	maxRows := 100
-	if badConversions > 0 {
-		l := conv.SampleBadRows(maxRows)
-		if int64(len(l)) < badConversions {
-			f.WriteString("A sample of rows that generated conversion errors:\n")
-		} else {
-			f.WriteString("Rows that generated conversion errors:\n")
-		}
-		for _, r := range l {
-			_, err := f.WriteString("  " + r + "\n")
-			if err != nil {
-				fmt.Fprintf(out, "Can't write out bad data file: %v\n", err)
-				return
-			}
-		}
-	}
-	if badWrites > 0 {
-		l := bw.SampleBadRows(maxRows)
-		if int64(len(l)) < badWrites {
-			f.WriteString("A sample of rows that successfully converted but couldn't be written to Spanner:\n")
-		} else {
-			f.WriteString("Rows that successfully converted but couldn't be written to Spanner:\n")
-		}
-		for _, r := range l {
-			_, err := f.WriteString("  " + r + "\n")
-			if err != nil {
-				fmt.Fprintf(out, "Can't write out bad data file: %v\n", err)
-				return
-			}
-		}
-	}
-	if badDataStreaming > 0 {
-		err = writeBadStreamingData(conv, f)
-		if err != nil {
-			fmt.Fprintf(out, "Can't write out bad data file: %v\n", err)
-			return
-		}
-	}
-
-	fmt.Fprintf(out, "See file '%s' for details of bad rows\n", name)
-}
-
-// getBadStreamingDataCount returns the total sum of bad and dropped records during
-// streaming migration process.
-func getBadStreamingDataCount(conv *internal.Conv) int64 {
-	badDataCount := int64(0)
-
-	for _, x := range conv.Audit.StreamingStats.BadRecords {
-		badDataCount += utils.SumMapValues(x)
-	}
-	for _, x := range conv.Audit.StreamingStats.DroppedRecords {
-		badDataCount += utils.SumMapValues(x)
-	}
-	return badDataCount
-}
-
-// writeBadStreamingData writes sample of bad records and dropped records during streaming
-// migration process to bad data file.
-func writeBadStreamingData(conv *internal.Conv, f *os.File) error {
-	f.WriteString("\nBad data encountered during streaming migration:\n\n")
-
-	stats := (conv.Audit.StreamingStats)
-
-	badRecords := int64(0)
-	for _, x := range stats.BadRecords {
-		badRecords += utils.SumMapValues(x)
-	}
-	droppedRecords := int64(0)
-	for _, x := range stats.DroppedRecords {
-		droppedRecords += utils.SumMapValues(x)
-	}
-
-	if badRecords > 0 {
-		l := stats.SampleBadRecords
-		if int64(len(l)) < badRecords {
-			f.WriteString("A sample of records that generated conversion errors:\n")
-		} else {
-			f.WriteString("Records that generated conversion errors:\n")
-		}
-		for _, r := range l {
-			_, err := f.WriteString("  " + r + "\n")
-			if err != nil {
-				return err
-			}
-		}
-		f.WriteString("\n")
-	}
-	if droppedRecords > 0 {
-		l := stats.SampleBadWrites
-		if int64(len(l)) < droppedRecords {
-			f.WriteString("A sample of records that successfully converted but couldn't be written to Spanner:\n")
-		} else {
-			f.WriteString("Records that successfully converted but couldn't be written to Spanner:\n")
-		}
-		for _, r := range l {
-			_, err := f.WriteString("  " + r + "\n")
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// ProcessDump invokes process dump function from a sql package based on driver selected.
-func ProcessDump(driver string, conv *internal.Conv, r *internal.Reader) error {
-	switch driver {
-	case constants.MYSQLDUMP:
-		return common.ProcessDbDump(conv, r, mysql.DbDumpImpl{})
-	case constants.PGDUMP:
-		return common.ProcessDbDump(conv, r, postgres.DbDumpImpl{})
-	default:
-		return fmt.Errorf("process dump for driver %s not supported", driver)
-	}
-}
-
-func GetInfoSchemaFromCloudSQL(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile) (common.InfoSchema, error) {
-	driver := sourceProfile.Driver
-	switch driver {
-	case constants.MYSQL:
-		d, err := cloudsqlconn.NewDialer(context.Background(), cloudsqlconn.WithIAMAuthN())
-		if err != nil {
-			return nil, fmt.Errorf("cloudsqlconn.NewDialer: %w", err)
-		}
-		var opts []cloudsqlconn.DialOption
-		instanceName := fmt.Sprintf("%s:%s:%s", sourceProfile.ConnCloudSQL.Mysql.Project, sourceProfile.ConnCloudSQL.Mysql.Region, sourceProfile.ConnCloudSQL.Mysql.InstanceName)
-		mysqldriver.RegisterDialContext("cloudsqlconn",
-			func(ctx context.Context, addr string) (net.Conn, error) {
-				return d.Dial(ctx, instanceName, opts...)
-			})
-
-		dbURI := fmt.Sprintf("%s:empty@cloudsqlconn(localhost:3306)/%s?parseTime=true",
-			sourceProfile.ConnCloudSQL.Mysql.User, sourceProfile.ConnCloudSQL.Mysql.Db)
-
-		db, err := sql.Open("mysql", dbURI)
-		if err != nil {
-			return nil, fmt.Errorf("sql.Open: %w", err)
-		}
-		return mysql.InfoSchemaImpl{
-			DbName:        sourceProfile.ConnCloudSQL.Mysql.Db,
-			Db:            db,
-			SourceProfile: sourceProfile,
-			TargetProfile: targetProfile,
-		}, nil
-	case constants.POSTGRES:
-		d, err := cloudsqlconn.NewDialer(context.Background(), cloudsqlconn.WithIAMAuthN())
-		if err != nil {
-			return nil, fmt.Errorf("cloudsqlconn.NewDialer: %w", err)
-		}
-		var opts []cloudsqlconn.DialOption
-
-		dsn := fmt.Sprintf("user=%s database=%s", sourceProfile.ConnCloudSQL.Pg.User, sourceProfile.ConnCloudSQL.Pg.Db)
-		config, err := pgx.ParseConfig(dsn)
-		if err != nil {
-			return nil, err
-		}
-		instanceName := fmt.Sprintf("%s:%s:%s", sourceProfile.ConnCloudSQL.Pg.Project, sourceProfile.ConnCloudSQL.Pg.Region, sourceProfile.ConnCloudSQL.Pg.InstanceName)
-		config.DialFunc = func(ctx context.Context, network, instance string) (net.Conn, error) {
-			return d.Dial(ctx, instanceName, opts...)
-		}
-		dbURI := stdlib.RegisterConnConfig(config)
-		db, err := sql.Open("pgx", dbURI)
-		if err != nil {
-			return nil, fmt.Errorf("sql.Open: %w", err)
-		}
-		temp := false
-		return postgres.InfoSchemaImpl{
-			Db:             db,
-			SourceProfile:  sourceProfile,
-			TargetProfile:  targetProfile,
-			IsSchemaUnique: &temp, //this is a workaround to set a bool pointer
-		}, nil
-	default:
-		return nil, fmt.Errorf("driver %s not supported", driver)
-	}
-}
-
-func GetInfoSchema(sourceProfile profiles.SourceProfile, targetProfile profiles.TargetProfile) (common.InfoSchema, error) {
-	connectionConfig, err := connectionConfig(sourceProfile)
-	if err != nil {
-		return nil, err
-	}
-	driver := sourceProfile.Driver
-	switch driver {
-	case constants.MYSQL:
-		db, err := sql.Open(driver, connectionConfig.(string))
-		dbName := getDbNameFromSQLConnectionStr(driver, connectionConfig.(string))
-		if err != nil {
-			return nil, err
-		}
-		return mysql.InfoSchemaImpl{
-			DbName:        dbName,
-			Db:            db,
-			SourceProfile: sourceProfile,
-			TargetProfile: targetProfile,
-		}, nil
-	case constants.POSTGRES:
-		db, err := sql.Open(driver, connectionConfig.(string))
-		if err != nil {
-			return nil, err
-		}
-		temp := false
-		return postgres.InfoSchemaImpl{
-			Db:             db,
-			SourceProfile:  sourceProfile,
-			TargetProfile:  targetProfile,
-			IsSchemaUnique: &temp, //this is a workaround to set a bool pointer
-		}, nil
-	case constants.DYNAMODB:
-		mySession := session.Must(session.NewSession())
-		dydbClient := dydb.New(mySession, connectionConfig.(*aws.Config))
-		var dydbStreamsClient *dynamodbstreams.DynamoDBStreams
-		if sourceProfile.Conn.Streaming {
-			newSession := session.Must(session.NewSession())
-			dydbStreamsClient = dynamodbstreams.New(newSession, connectionConfig.(*aws.Config))
-		}
-		return dynamodb.InfoSchemaImpl{
-			DynamoClient:        dydbClient,
-			SampleSize:          profiles.GetSchemaSampleSize(sourceProfile),
-			DynamoStreamsClient: dydbStreamsClient,
-		}, nil
-	case constants.SQLSERVER:
-		db, err := sql.Open(driver, connectionConfig.(string))
-		dbName := getDbNameFromSQLConnectionStr(driver, connectionConfig.(string))
-		if err != nil {
-			return nil, err
-		}
-		return sqlserver.InfoSchemaImpl{DbName: dbName, Db: db}, nil
-	case constants.ORACLE:
-		db, err := sql.Open(driver, connectionConfig.(string))
-		dbName := getDbNameFromSQLConnectionStr(driver, connectionConfig.(string))
-		if err != nil {
-			return nil, err
-		}
-		return oracle.InfoSchemaImpl{DbName: strings.ToUpper(dbName), Db: db, SourceProfile: sourceProfile, TargetProfile: targetProfile}, nil
-	default:
-		return nil, fmt.Errorf("driver %s not supported", driver)
-	}
 }
