+++ sources/dynamodb/toddl_test.go
@@ -58,7 +58,8 @@ func TestToSpannerType(t *testing.T) {
 	}
 	conv.SrcSchema[name] = srcSchema
 	conv.Audit = audit
-	assert.Nil(t, common.SchemaToSpannerDDL(conv, ToDdlImpl{}))
+	schemaToSpanner := common.SchemaToSpannerImpl{}
+	assert.Nil(t, schemaToSpanner.SchemaToSpannerDDL(conv, ToDdlImpl{}))
 	actual := conv.SpSchema[name]
 	dropComments(&actual) // Don't test comment.
 	expected := ddl.CreateTable{
@@ -118,7 +119,8 @@ func TestToSpannerPostgreSQLDialectType(t *testing.T) {
 	}
 	conv.SrcSchema["t1"] = srcSchema
 	conv.Audit = audit
-	assert.Nil(t, common.SchemaToSpannerDDL(conv, ToDdlImpl{}))
+	schemaToSpanner := common.SchemaToSpannerImpl{}
+	assert.Nil(t, schemaToSpanner.SchemaToSpannerDDL(conv, ToDdlImpl{}))
 	actual := conv.SpSchema["t1"]
 	dropComments(&actual) // Don't test comment.
 	expected := ddl.CreateTable{
