+++ sources/postgres/toddl_test.go
@@ -162,7 +162,8 @@ func TestToSpannerType(t *testing.T) {
 		PrimaryKeys: []schema.Key{{ColId: "c10"}},
 	}
 	conv.UsedNames = map[string]bool{"ref_table": true, "ref_table2": true}
-	assert.Nil(t, common.SchemaToSpannerDDL(conv, ToDdlImpl{}))
+	schemaToSpanner := common.SchemaToSpannerImpl{}
+	assert.Nil(t, schemaToSpanner.SchemaToSpannerDDL(conv, ToDdlImpl{}))
 	actual := conv.SpSchema[tableId]
 	dropComments(&actual) // Don't test comment.
 	expected := ddl.CreateTable{
@@ -242,7 +243,8 @@ func TestToExperimentalSpannerType(t *testing.T) {
 		PrimaryKeys: []schema.Key{{ColId: "c10"}},
 	}
 	conv.UsedNames = map[string]bool{"ref_table": true, "ref_table2": true}
-	assert.Nil(t, common.SchemaToSpannerDDL(conv, ToDdlImpl{}))
+	schemaToSpanner := common.SchemaToSpannerImpl{}
+	assert.Nil(t, schemaToSpanner.SchemaToSpannerDDL(conv, ToDdlImpl{}))
 	actual := conv.SpSchema[tableId]
 	dropComments(&actual) // Don't test comment.
 	expected := ddl.CreateTable{
