+++ sources/sqlserver/toddl_test.go
@@ -164,7 +164,8 @@ func TestToSpannerType(t *testing.T) {
 		PrimaryKeys: []schema.Key{{ColId: "c19"}},
 	}
 	conv.UsedNames = map[string]bool{"ref_table": true, "ref_table2": true}
-	assert.Nil(t, common.SchemaToSpannerDDL(conv, ToDdlImpl{}))
+	schemaToSpanner := common.SchemaToSpannerImpl{}
+	assert.Nil(t, schemaToSpanner.SchemaToSpannerDDL(conv, ToDdlImpl{}))
 	actual := conv.SpSchema[tableId]
 	dropComments(&actual) // Don't test comment.
 	expected := ddl.CreateTable{
@@ -262,7 +263,8 @@ func TestToSpannerPostgreSQLDialectType(t *testing.T) {
 		PrimaryKeys: []schema.Key{{ColId: "c19"}},
 	}
 	conv.UsedNames = map[string]bool{"ref_table": true, "ref_table2": true}
-	assert.Nil(t, common.SchemaToSpannerDDL(conv, ToDdlImpl{}))
+	schemaToSpanner := common.SchemaToSpannerImpl{}
+	assert.Nil(t, schemaToSpanner.SchemaToSpannerDDL(conv, ToDdlImpl{}))
 	actual := conv.SpSchema[tableId]
 	dropComments(&actual) // Don't test comment.
 	expected := ddl.CreateTable{
